apps/back_admin/db_management: skip 'alembic_version' table

+++ apps/back_admin/service/db_management/db_eraser.py
@@ -8,6 +8,7 @@ async def clear_database_data(session: AsyncSession):
     tables_descriptor = (await session.execute(text("SHOW TABLES"))).scalars()
 
     for table in tables_descriptor:
-        await session.execute(text(f"DELETE FROM `{table}`"))
+        if table != "alembic_version":
+            await session.execute(text(f"DELETE FROM `{table}`"))
 
     await session.execute(text("SET FOREIGN_KEY_CHECKS = 1"))
